Move recursion protection test to where it matters
The EntityValueFactory only looks at a single property at a time, so testing it here, does not show that it works when looking at an entire POCO.
